Implement OOP for Parser
