+++ src/main/java/Ui.java
@@ -1,2 +1,9 @@
-package PACKAGE_NAME;public class Ui {
+public class Ui {
+    // greeting message
+    String greetings = "    Hi there! 👋 I'm Duke\n"
+            + "    What can I do for you?";
+
+    // divider
+    String lines = "    ---------------------------------";
+
 }
